+++ crates/factory_sim/tests/performance_budgets.rs
@@ -308,8 +308,8 @@ fn place_entities(
 
 fn benchmark_placement_request(
     prototype_id: factory_data::EntityPrototypeId,
-    x: i32,
-    y: i32,
+    x: i64,
+    y: i64,
     direction: Direction,
 ) -> factory_sim::placement::EntityPlacementRequest {
     factory_sim::placement::EntityPlacementRequest {
@@ -437,7 +437,7 @@ fn place_fluid_fixtures(sim: &mut Simulation, count: usize) {
     panic!("could only place {placed} of {count} fluid benchmark fixtures");
 }
 
-fn deterministic_tile_coords(sim: &Simulation) -> Vec<(i32, i32)> {
+fn deterministic_tile_coords(sim: &Simulation) -> Vec<(i64, i64)> {
     let mut chunks = sim.world().chunks.keys().copied().collect::<Vec<_>>();
     chunks.sort_unstable();
     chunks
